api/requirements.txt
pillow
numpy
requests 